version 4

+++ modules/task_4/bakaeva_m_convex_hull_graham/convex_hull_graham.cpp
@@ -149,7 +149,7 @@ std::vector<std::pair<double, double> > grahamScanParallel(
     });
 
     // Wait for threads to complete
-    for (size_t i = 0; i < numberOfThreads; i++) {
+    for (int i = 0; i < numberOfThreads; i++) {
         threadsPull[i].join();
     }
